Add RESTCode enum, update interfaces and methods
A new enum `RESTCode` was added in `RESTCode.cs` to represent various REST response codes with descriptions. The `State` enum in `State.cs` had an unused `using System.Reflection;` directive removed. A new method `GetEventMemberList` was added to the `IDatabaseService` interface in `IDatabaseService.cs`. The `IUserService` interface in `IUserService.cs` was updated to change the return type of `LoginUser` from `bool` to `RESTCode`.

In `DatabaseService.cs`, the `InsertToInvite` and `InsertToEventMember` methods were implemented to add entities to the database and save changes. A new method `GetEventMemberList` was implemented in `DatabaseService.cs` to retrieve a list of event members based on an event. In `UserService.cs`, the `LoginUser` method was updated to return `RESTCode` values instead of `bool` to provide more detailed response codes.

The `LoginController` in `LoginController.cs` was updated to handle the new `RESTCode` return values from `LoginUser` and return appropriate HTTP responses based on the `RESTCode`.